feat: Make page links reusable and update on new pages
